chore: :truck: moves auth service to correct folder
